requirements.txt
Flask-SQLAlchemy==2.3.2
Jinja2==2.10
numpy==1.14.3 
scipy==1.1.0
matplotlib
seaborn 